Fixed relationship from integer to ubigInt

+++ database/migrations/2024_07_24_114100_create_add_trx_ref.php
@@ -13,7 +13,7 @@ return new class extends Migration
     {
         Schema::table('payments', function (Blueprint $table) {
             $table->string('trx_ref')->nullable()->after('user_id');
-            $table->integer('license_id')->nullable()->after('trx_ref');
+            $table->unsignedBigInteger('license_id')->nullable()->after('trx_ref');
 
             $table->foreign('license_id')->references('id')->on('licenses')->onDelete('cascade');
         });
